Move feature_statistic.py into statics.py
