+++ examples/federatedml-1.x-examples/hetero_feature_selection/test_hetero_feature_selection_statistic_model_conf.json
@@ -1,0 +1,124 @@
+{
+    "initiator": {
+        "role": "guest",
+        "party_id": 10000
+    },
+    "job_parameters": {
+        "work_mode": 0
+    },
+    "role": {
+        "guest": [
+            10000
+        ],
+        "host": [
+            10000
+        ]
+    },
+    "role_parameters": {
+        "guest": {
+            "args": {
+                "data": {
+                    "train_data": [
+                        {
+                            "name": "breast_hetero_guest",
+                            "namespace": "experiment"
+                        }
+                    ]
+                }
+            },
+            "dataio_0": {
+                "with_label": [
+                    true
+                ],
+                "label_name": [
+                    "y"
+                ],
+                "label_type": [
+                    "int"
+                ],
+                "output_format": [
+                    "dense"
+                ]
+            }
+        },
+        "host": {
+            "args": {
+                "data": {
+                    "train_data": [
+                        {
+                            "name": "breast_hetero_host",
+                            "namespace": "experiment"
+                        }
+                    ]
+                }
+            },
+            "dataio_0": {
+                "with_label": [
+                    false
+                ],
+                "output_format": [
+                    "dense"
+                ]
+            }
+        }
+    },
+    "algorithm_parameters": {
+        "hetero_feature_binning_0": {
+            "method": "quantile",
+            "compress_thres": 10000,
+            "head_size": 10000,
+            "error": 0.001,
+            "bin_num": 10,
+            "bin_indexes": -1,
+            "bin_names": null,
+            "category_indexes": null,
+            "category_names": null,
+            "adjustment_factor": 0.5,
+            "local_only": false,
+            "transform_param": {
+                "transform_cols": -1,
+                "transform_names": null,
+                "transform_type": "bin_num"
+            }
+        },
+        "statistic_0": {
+            "statistics": "summary",
+            "column_indexes": -1,
+            "column_names": []
+        },
+        "psi_0": {
+            "max_bin_num": 20
+        },
+
+        "hetero_feature_selection_0": {
+            "select_col_indexes": -1,
+            "select_names": [],
+            "filter_methods": [
+                "iv_filter",
+                "statistic_filter",
+                "psi_filter"
+            ],
+            "iv_param": {
+                "metrics": ["iv", "iv", "iv"],
+                "filter_type": ["threshold", "top_k", "top_percentile"],
+                "take_high": true,
+                "threshold": [0.03, 15, 0.7],
+                "host_thresholds": [[0.15], null, null],
+                "select_federated": true
+            },
+            "statistic_param": {
+                "metrics": ["skewness", "skewness", "kurtosis",
+                    "median", "missing_count", "missing_ratio"],
+                "filter_type": "threshold",
+                "take_high": [true, false, true, true, false, false],
+                "threshold": [-10, 10, -1.5, -1.5, 1, 0.1]
+            },
+            "psi_param": {
+                "metrics": "psi",
+                "filter_type": "threshold",
+                "take_high": false,
+                "threshold": -0.1
+            }
+        }
+    }
+}
