+++ examples/federatedml-1.x-examples/hetero_feature_selection/test_hetero_feature_statistic_model_selection_job_dsl.json
@@ -1,0 +1,96 @@
+{
+    "components" : {
+        "dataio_0": {
+            "module": "DataIO",
+            "input": {
+                "data": {
+                    "data": [
+                        "args.train_data"
+                    ]
+                }
+            },
+            "output": {
+                "data": ["train_data"],
+                "model": ["dataio"]
+            },
+			"need_deploy": true
+         },
+        "intersection_0": {
+             "module": "Intersection",
+             "input": {
+                 "data": {
+                     "data": [
+                         "dataio_0.train_data"
+                     ]
+                 }
+             },
+             "output": {
+                 "data": ["train_data"]
+             }
+        },
+        "hetero_feature_binning_0": {
+            "module": "HeteroFeatureBinning",
+            "input": {
+                "data": {
+                    "data": [
+                        "intersection_0.train_data"
+                    ]
+                }
+            },
+            "output": {
+                "data": ["train_data"],
+                "model": ["model"]
+            }
+        },
+        "statistic_0": {
+            "module": "DataStatistics",
+            "input": {
+                "data": {
+                    "data": [
+                        "intersection_0.train_data"
+                    ]
+                }
+            },
+            "output": {
+                "data": ["train_data"],
+                "model": ["model"]
+            }
+        },
+        "psi_0": {
+            "module": "PSI",
+            "input": {
+                "data": {
+                    "train_data": [
+                        "intersection_0.train_data"
+                    ],
+                    "eval_data": [
+                        "intersection_0.train_data"
+                    ]
+                }
+            },
+            "output": {
+                "data": ["train_data"],
+                "model": ["model"]
+            }
+        },
+        "hetero_feature_selection_0": {
+            "module": "HeteroFeatureSelection",
+            "input": {
+                "data": {
+                    "data": [
+                        "intersection_0.train_data"
+                    ]
+                },
+                "isometric_model": [
+                    "hetero_feature_binning_0.model",
+                    "statistic_0.model",
+                    "psi_0.model"
+                ]
+            },
+            "output": {
+                "data": ["train_data"],
+                "model": ["model"]
+            }
+        }
+    }
+}
